Inventory now properly changes player weight
We are handling the changing of inventory weight to just change the player's actual weight as we can use a base volume and adjust volume from there depending on how much the player is carrying. The SetWeight() will be moved to the C++ file when this occurs but for now, this will do.