Changed to full mesh for First Person
Previously, the mesh was just floating arms. Changed to a full mesh which required changing the camera position, and adding a spring arm (attached to the head bone) so that there was some head bobbing, but not too much. Will eventually have a setting that eliminates head bobbing entirely (for player choice).